Safari-611.1.10/LayoutTests/fast/css-grid-layout/grid-track-sizing-with-percentages-and-orthogonal-flows.html

<!DOCTYPE html>
<link href="resources/grid.css" rel="stylesheet">
<link href="resources/grid-alignment.css" rel="stylesheet">
<link href="../css-intrinsic-dimensions/resources/width-keyword-classes.css" rel="stylesheet">
<link href="../css-intrinsic-dimensions/resources/height-keyword-classes.css" rel="stylesheet">
<style>
body {
    margin: 0;
}

.container {
    position: relative;
}

.grid {
    font: 10px/1 Ahem;
    grid-template-columns: 100px 25%;
    grid-template-rows: 50% 150px;
}

.definiteGridWithPercentageGaps {
    grid-template: 5px 5px / auto;
    grid-row-gap: 80%;
    grid-column-gap: 20%;
    height: 200px;
    width: 100px;
    justify-content: start;
    align-content: start;
}
</style>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<body onload="checkLayout('.grid')">
<div id="log"></div>
<p>This test checks that grid tracks are correctly sized when using percentage lengths and orthogonal flows.</p>

<p>HORIZONTAL-TB container with VERTICAL-LR items.</p>
<div class="container">
    <div class="grid itemsStart contentStart fit-content" data-expected-width="120" data-expected-height="260">
        <div class="firstRowFirstColumn   verticalLR" data-offset-x="0"  data-offset-y="0"  data-expected-width="10" data-expected-height="50">X X X</div>
        <div class="firstRowSecondColumn  verticalLR" data-offset-x="100" data-offset-y="0"  data-expected-width="10" data-expected-height="110">X X X X X X</div>
        <div class="secondRowFirstColumn  verticalLR" data-offset-x="0"  data-offset-y="130" data-expected-width="10" data-expected-height="110">X X X X X X</div>
        <div class="secondRowSecondColumn verticalLR" data-offset-x="100" data-offset-y="130" data-expected-width="20" data-expected-height="150">X X X X X X X X X X X X</div>
    </div>
</div>

<p>HORIZONTAL-TB container with VERTICAL-RL items.</p>
<div class="container">
    <div class="grid itemsStart contentStart fit-content" data-expected-width="120" data-expected-height="260">
        <div class="firstRowFirstColumn   verticalRL" data-offset-x="0"  data-offset-y="0"  data-expected-width="10" data-expected-height="50">X X X</div>
        <div class="firstRowSecondColumn  verticalRL" data-offset-x="100" data-offset-y="0"  data-expected-width="10" data-expected-height="110">X X X X X X</div>
        <div class="secondRowFirstColumn  verticalRL" data-offset-x="0"  data-offset-y="130" data-expected-width="10" data-expected-height="110">X X X X X X</div>
        <div class="secondRowSecondColumn verticalRL" data-offset-x="100" data-offset-y="130" data-expected-width="20" data-expected-height="150">X X X X X X X X X X X X</div>
    </div>
</div>

<p>VERTICAL-LR container with HORIZONTAL-TB items.</p>
<div class="container">
    <div class="grid itemsStart contentStart verticalLR fit-content" data-expected-width="260" data-expected-height="120">
        <div class="firstRowFirstColumn   horizontalTB" data-offset-x="0"   data-offset-y="0"  data-expected-width="50" data-expected-height="10">X X X</div>
        <div class="firstRowSecondColumn  horizontalTB" data-offset-x="0" data-offset-y="100" data-expected-width="110" data-expected-height="10">X X X X X X</div>
        <div class="secondRowFirstColumn  horizontalTB" data-offset-x="130"   data-offset-y="0"  data-expected-width="110" data-expected-height="10">X X X X X X</div>
        <div class="secondRowSecondColumn horizontalTB" data-offset-x="130" data-offset-y="100" data-expected-width="150" data-expected-height="20">X X X X X X X X X X X X</div>
    </div>
</div>

<p>VERTICAL-RL container with HORIZONTAL-TB items.</p>
<div class="container">
    <div class="grid itemsStart contentStart verticalRL fit-content" data-expected-width="260" data-expected-height="120">
        <div class="firstRowFirstColumn   horizontalTB" data-offset-x="210"   data-offset-y="0"  data-expected-width="50" data-expected-height="10">X X X</div>
        <div class="firstRowSecondColumn  horizontalTB" data-offset-x="150" data-offset-y="100" data-expected-width="110" data-expected-height="10">X X X X X X</div>
        <div class="secondRowFirstColumn  horizontalTB" data-offset-x="20"   data-offset-y="0"  data-expected-width="110" data-expected-height="10">X X X X X X</div>
        <div class="secondRowSecondColumn horizontalTB" data-offset-x="-20" data-offset-y="100" data-expected-width="150" data-expected-height="20">X X X X X X X X X X X X</div>
    </div>
</div>

<p>HORIZONTAL-TB container with VERTICAL-LR item with percentage gaps.</p>
<div class="grid definiteGridWithPercentageGaps" data-expected-width="100" data-expected-height="200">
    <div class="bothRowFirstColumn verticalLR" style="background: cyan;" data-expected-width="10" data-expected-height="170">XXXX XXXXX</div>
</div>

</body>
